Fixed how protoc plugins are applied
