+++ gradle-plugin/api/gradle-plugin.api
@@ -39,11 +39,7 @@ public class kotlinx/rpc/RpcStrictModeExtension {
 
 public final class kotlinx/rpc/VersionsKt {
 	public static final field BUF_TOOL_VERSION Ljava/lang/String;
-	public static final field GRPC_KOTLIN_VERSION Ljava/lang/String;
-	public static final field GRPC_VERSION Ljava/lang/String;
 	public static final field LIBRARY_VERSION Ljava/lang/String;
-	public static final field PLUGIN_VERSION Ljava/lang/String;
-	public static final field PROTOBUF_VERSION Ljava/lang/String;
 }
 
 public class kotlinx/rpc/buf/BufExtension {
@@ -179,17 +175,15 @@ public final class kotlinx/rpc/protoc/ProcessProtoFilesKt$inlined$sam$i$org_grad
 
 public abstract interface class kotlinx/rpc/protoc/ProtoSourceSet {
 	public abstract fun getName ()Ljava/lang/String;
+	public abstract fun getPlugins ()Lorg/gradle/api/NamedDomainObjectContainer;
 	public abstract fun getProto ()Lorg/gradle/api/file/SourceDirectorySet;
+	public abstract fun plugins (Lorg/gradle/api/Action;)V
 	public fun proto (Lorg/gradle/api/Action;)V
-	public abstract fun protocPlugin (Lkotlinx/rpc/protoc/ProtocPlugin;)V
-	public abstract fun protocPlugin (Lorg/gradle/api/NamedDomainObjectProvider;)V
 }
 
 public abstract interface class kotlinx/rpc/protoc/ProtocExtension {
 	public abstract fun buf (Lorg/gradle/api/Action;)V
 	public abstract fun getBuf ()Lkotlinx/rpc/buf/BufExtension;
-	public abstract fun getPlugins ()Lorg/gradle/api/NamedDomainObjectContainer;
-	public abstract fun plugins (Lorg/gradle/api/Action;)V
 }
 
 public class kotlinx/rpc/protoc/ProtocPlugin {
