+++ gradle-plugin/src/test/resources/projects/GrpcJvmProjectTest/custom_protoc_plugins_must_declare_an_artifact/build.gradle.kts
@@ -11,15 +11,13 @@ plugins {
 }
 
 rpc {
-    protoc {
-        plugins {
-            create("myPlugin") {}
-        }
-    }
+    protoc()
 }
 
 protoSourceSets {
     main {
-        protocPlugin(rpc.protoc.plugins.named("myPlugin"))
+        plugins {
+            create("myPlugin") {}
+        }
     }
 }
